+++ CodeTester-IntelliJ-IDEA-UITest/src/test/kotlin/de/fxnm/fixtures/IdeaFrame.kt
@@ -55,6 +55,9 @@ class IdeaFrame(remoteRobot: RemoteRobot, remoteComponent: RemoteComponent) :
                 findAll<ComponentFixture>(byXpath("//div[@class='JProgressBar']")).isEmpty()
             }
         }
+
+        // Force UI update
+        rightClick()
     }
 
     private fun isDumbMode(): Boolean = callJs("""
@@ -65,7 +68,7 @@ class IdeaFrame(remoteRobot: RemoteRobot, remoteComponent: RemoteComponent) :
 
     fun showCodeTesterExplorer() {
         // FIXME: 18.01.2021 BUG IF CURRENT OPEN TOOL WINDOW IS NOT THE MAIN TOOL WINDOW
-        if (findAll<CodeTesterExplorer>(CodeTesterExplorer::class.java).isEmpty()) {
+        if (findAll(CodeTesterExplorer::class.java).isEmpty()) {
             find(ComponentFixture::class.java,
                  byXpath("//div[@accessiblename='CodeTester' and @class='StripeButton' and @text='CodeTester']")).click()
         }